requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
azure-identity==1.21.0
azure-storage-blob==12.25.1
Office365_REST_Python_Client==2.6.2
PyPDF2==3.0.1
python-dotenv==1.1.0
pandas==2.2.3
pillow==11.3.0
pdf2image==1.17.0
pytesseract==0.3.13